fixed some red lines, now need to change main class and console class, to be able to add suppliers and all that stuff(trucks,branches,products, etc...)

+++ dev/src/Domain/Product.java
@@ -5,4 +5,4 @@ package Domain;
  * @param weight The weight in Kilograms (Kg). Must be a positive value.
  * */
 
-public record Product(String name, int weight) {}
+public record Product(String name, double weight) {}
